pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "brightness-monitor"
version = "0.1.0"
description = "Sync MacBook keyboard brightness to Claude API usage limits"
requires-python = ">=3.10"
dependencies = [
    "pyobjc-core>=10.0",
    "PyYAML>=6.0",
]

[project.scripts]
brightness-monitor = "brightness_monitor.main:main"
